+++ apscheduler/schedulers/blocking.py
@@ -1,7 +1,9 @@
 from __future__ import absolute_import
+
 from threading import Event
 
-from apscheduler.schedulers.base import BaseScheduler
+from apscheduler.schedulers.base import BaseScheduler, SchedulerState
+from apscheduler.util import TIMEOUT_MAX
 
 
 class BlockingScheduler(BaseScheduler):
@@ -9,14 +11,11 @@ class BlockingScheduler(BaseScheduler):
     A scheduler that runs in the foreground
     (:meth:`~apscheduler.schedulers.base.BaseScheduler.start` will block).
     """
-
-    MAX_WAIT_TIME = 4294967  # Maximum value accepted by Event.wait() on Windows
-
     _event = None
 
-    def start(self):
+    def start(self, *args, **kwargs):
         self._event = Event()
-        super(BlockingScheduler, self).start()
+        super(BlockingScheduler, self).start(*args, **kwargs)
         self._main_loop()
 
     def shutdown(self, wait=True):
@@ -24,9 +23,9 @@ class BlockingScheduler(BaseScheduler):
         self._event.set()
 
     def _main_loop(self):
-        wait_seconds = self.MAX_WAIT_TIME
-        while self.running:
-            self._event.wait(wait_seconds if wait_seconds is not None else self.MAX_WAIT_TIME)
+        wait_seconds = TIMEOUT_MAX
+        while self.state is not SchedulerState.stopped:
+            self._event.wait(wait_seconds)
             self._event.clear()
             wait_seconds = self._process_jobs()